Started refactoring MeasurementsScreen to use MVVM pattern
