+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/Koin.kt
@@ -4,10 +4,12 @@ import org.koin.core.KoinApplication
 import org.koin.core.context.startKoin
 import org.koin.core.module.Module
 import org.koin.dsl.module
-import org.noiseplanet.noisecapture.measurements.MeasurementService
+import org.noiseplanet.noisecapture.measurements.DefaultMeasurementService
+import org.noiseplanet.noisecapture.measurements.MeasurementsService
 import org.noiseplanet.noisecapture.permission.defaultPermissionModule
 import org.noiseplanet.noisecapture.permission.platformPermissionModule
 import org.noiseplanet.noisecapture.ui.features.home.homeModule
+import org.noiseplanet.noisecapture.ui.features.measurement.measurementModule
 import org.noiseplanet.noisecapture.ui.features.permission.requestPermissionModule
 
 /**
@@ -23,7 +25,9 @@ fun initKoin(
             },
 
             module {
-                single<MeasurementService> { MeasurementService(audioSource = get()) }
+                single<MeasurementsService> {
+                    DefaultMeasurementService(audioSource = get(), logger = get())
+                }
             },
 
             defaultPermissionModule,
@@ -31,6 +35,7 @@ fun initKoin(
 
             homeModule,
             requestPermissionModule,
+            measurementModule,
         )
         createEagerInstances()
     }
