+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/measurements/MeasurementsService.kt
@@ -1,128 +1,177 @@
 package org.noiseplanet.noisecapture.measurements
 
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.channels.awaitClose
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
+import org.koin.core.component.KoinComponent
 import org.noiseplanet.noisecapture.audio.AcousticIndicatorsData
 import org.noiseplanet.noisecapture.audio.AcousticIndicatorsProcessing
 import org.noiseplanet.noisecapture.audio.AudioSource
-import org.noiseplanet.noisecapture.audio.signal.SpectrumData
-import org.noiseplanet.noisecapture.audio.signal.WindowAnalysis
-import kotlin.properties.Delegates
-import kotlin.reflect.KProperty
+import org.noiseplanet.noisecapture.audio.WINDOW_TIME
+import org.noiseplanet.noisecapture.audio.signal.FAST_DECAY_RATE
+import org.noiseplanet.noisecapture.audio.signal.LevelDisplayWeightedDecay
+import org.noiseplanet.noisecapture.audio.signal.window.SpectrumData
+import org.noiseplanet.noisecapture.audio.signal.window.SpectrumDataProcessing
+import org.noiseplanet.noisecapture.log.Logger
 
-typealias AcousticIndicatorsCallback = (acousticIndicatorsData: AcousticIndicatorsData) -> Unit
-typealias SpectrumDataCallback = (spectrumData: SpectrumData) -> Unit
+/**
+ * Record, observe and save audio measurements.
+ */
+interface MeasurementsService {
 
-const val FFT_SIZE = 4096
-const val FFT_HOP = 2048
+    /**
+     * Starts recording audio through the provided audio source.
+     * If already a recording is already running, calling this again will have no effect.
+     */
+    fun startRecordingAudio()
 
-class MeasurementService(private val audioSource: AudioSource) {
+    /**
+     * Stops the currently running audio recording.
+     * If no recording is running, this will have no effect
+     */
+    fun stopRecordingAudio()
 
-    var storageObservers =
-        mutableListOf<(property: KProperty<*>, oldValue: Boolean, newValue: Boolean) -> Unit>()
+    /**
+     * Get a [Flow] of [AcousticIndicatorsData] from the currently running recording.
+     */
+    fun getAcousticIndicatorsFlow(): Flow<AcousticIndicatorsData>
 
-    private var storageActivated: Boolean by Delegates.observable(false) { property, oldValue, newValue ->
-        storageObservers.forEach {
-            it(property, oldValue, newValue)
-        }
+    /**
+     * Get a [Flow] of sound pressure level values.
+     */
+    fun getWeightedLeqFlow(): Flow<Double>
+
+    /**
+     * Get a [Flow] of sound pressure levels weighted by frequency band.
+     */
+    fun getWeightedSoundPressureLevelFlow(): Flow<DoubleArray>
+
+    /**
+     * Get a [Flow] of [SpectrumData] from the currently running recording.
+     */
+    fun getSpectrumDataFlow(): Flow<SpectrumData>
+}
+
+/**
+ * Default [MeasurementsService] implementation.
+ * Can be overridden in platforms to add specific behaviour.
+ */
+class DefaultMeasurementService(
+    private val audioSource: AudioSource,
+    private val logger: Logger,
+) : MeasurementsService, KoinComponent {
+
+    companion object {
+
+        const val FFT_SIZE = 4096
+        const val FFT_HOP = 2048
+
+        private const val SPL_DECAY_RATE = FAST_DECAY_RATE
+        private const val SPL_WINDOW_TIME = WINDOW_TIME
     }
-    private var acousticIndicatorsProcessing: AcousticIndicatorsProcessing? = null
-    private var fftTool: WindowAnalysis? = null
-    private var onAcousticIndicatorsData: AcousticIndicatorsCallback? = null
-    private var onSpectrumData: SpectrumDataCallback? = null
-    private var audioJob: Job? = null
 
-    @OptIn(DelicateCoroutinesApi::class)
-    private fun startAudioRecord() {
+    private var indicatorsProcessing: AcousticIndicatorsProcessing? = null
+    private var spectrumDataProcessing: SpectrumDataProcessing? = null
+
+    private var audioJob: Job? = null
+    private val acousticIndicatorsFlow = MutableSharedFlow<AcousticIndicatorsData>(
+        replay = 1,
+        onBufferOverflow = BufferOverflow.DROP_OLDEST
+    )
+    private val spectrumDataFlow = MutableSharedFlow<SpectrumData>(
+        replay = 1,
+        onBufferOverflow = BufferOverflow.DROP_OLDEST
+    )
+    private val coroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
+
+    override fun startRecordingAudio() {
         if (audioJob?.isActive == true) {
+            logger.debug("Audio recording is already running. Don't start again.")
             return
         }
-        audioJob = GlobalScope.launch {
-            audioSource.setup().collect { audioSamples ->
-                if (onSpectrumData != null) {
-                    if (fftTool == null) {
-                        fftTool = WindowAnalysis(audioSamples.sampleRate, FFT_SIZE, FFT_HOP)
+        logger.debug("Starting recording audio samples...")
+        // Start recording and processing audio samples in a background thread
+        audioJob = coroutineScope.launch {
+            audioSource.setup()
+                .flowOn(Dispatchers.Default)
+                .collect { audioSamples ->
+                    // Process acoustic indicators
+                    if (indicatorsProcessing?.sampleRate != audioSamples.sampleRate) {
+                        logger.debug("Processing audio indicators with sample rate of ${audioSamples.sampleRate}")
+                        indicatorsProcessing = AcousticIndicatorsProcessing(audioSamples.sampleRate)
                     }
-                    fftTool?.pushSamples(audioSamples.epoch, audioSamples.samples)
-                        ?.forEach { spectrumData ->
-                            onSpectrumData?.let { callback -> callback(spectrumData) }
+                    indicatorsProcessing?.processSamples(audioSamples)
+                        ?.forEach {
+                            acousticIndicatorsFlow.tryEmit(it)
                         }
-                }
-                if (onAcousticIndicatorsData != null || storageActivated) {
-                    if (acousticIndicatorsProcessing == null) {
-                        acousticIndicatorsProcessing = AcousticIndicatorsProcessing(
-                            audioSamples.sampleRate
+
+                    // Process spectrum data
+                    if (spectrumDataProcessing?.sampleRate != audioSamples.sampleRate) {
+                        logger.debug("Processing spectrum data with sample rate of ${audioSamples.sampleRate}")
+                        spectrumDataProcessing = SpectrumDataProcessing(
+                            sampleRate = audioSamples.sampleRate,
+                            windowSize = FFT_SIZE,
+                            windowHop = FFT_HOP
                         )
                     }
-                    acousticIndicatorsProcessing!!.processSamples(audioSamples)
-                        .forEach { acousticIndicators ->
-                            if (onAcousticIndicatorsData != null) {
-                                onAcousticIndicatorsData?.let { callback ->
-                                    callback(
-                                        acousticIndicators
-                                    )
-                                }
-                            }
+                    spectrumDataProcessing?.pushSamples(audioSamples.epoch, audioSamples.samples)
+                        ?.forEach {
+                            spectrumDataFlow.tryEmit(it)
                         }
+
                 }
-            }
         }
     }
 
-    private fun stopAudioRecord() {
+    override fun stopRecordingAudio() {
+        if (audioJob == null || audioJob?.isActive == false) {
+            logger.debug("Audio recording is already stopped. Don't stop again.")
+            return
+        }
         audioJob?.cancel()
         audioSource.release()
     }
 
-    /**
-     * Start collecting measurements to be forwarded to observers
-     */
-    fun collectAudioIndicators(): Flow<AcousticIndicatorsData> = callbackFlow {
-        setAudioIndicatorsObserver { trySend(it) }
-        awaitClose {
-            resetAudioIndicatorsObserver()
-        }
+    override fun getAcousticIndicatorsFlow(): Flow<AcousticIndicatorsData> {
+        return acousticIndicatorsFlow.asSharedFlow()
     }
 
-    fun collectSpectrumData(): Flow<SpectrumData> = callbackFlow {
-        setSpectrumDataObserver { trySend(it) }
-        awaitClose {
-            resetSpectrumDataObserver()
-        }
+    override fun getSpectrumDataFlow(): Flow<SpectrumData> {
+        return spectrumDataFlow.asSharedFlow()
     }
 
-    private fun setSpectrumDataObserver(onSpectrumData: SpectrumDataCallback) {
-        this.onSpectrumData = onSpectrumData
-        startAudioRecord()
+    override fun getWeightedLeqFlow(): Flow<Double> {
+        val levelDisplay = LevelDisplayWeightedDecay(SPL_DECAY_RATE, SPL_WINDOW_TIME)
+
+        return getAcousticIndicatorsFlow()
+            .map {
+                levelDisplay.getWeightedValue(it.leq)
+            }
     }
 
-    private fun canReleaseAudio(): Boolean {
-        return !storageActivated &&
-            onAcousticIndicatorsData == null &&
-            onAcousticIndicatorsData == null
-    }
+    override fun getWeightedSoundPressureLevelFlow(): Flow<DoubleArray> {
+        var levelDisplayBands: Array<LevelDisplayWeightedDecay>? = null
 
-    private fun resetSpectrumDataObserver() {
-        onSpectrumData = null
-        if (canReleaseAudio()) {
-            stopAudioRecord()
-        }
-    }
-
-    private fun setAudioIndicatorsObserver(onAcousticIndicatorsData: AcousticIndicatorsCallback) {
-        startAudioRecord()
-        this.onAcousticIndicatorsData = onAcousticIndicatorsData
-    }
-
-    private fun resetAudioIndicatorsObserver() {
-        onAcousticIndicatorsData = null
-        if (canReleaseAudio()) {
-            stopAudioRecord()
-        }
+        return getAcousticIndicatorsFlow()
+            .map { indicators ->
+                if (levelDisplayBands == null) {
+                    levelDisplayBands = Array(indicators.nominalFrequencies.size) {
+                        LevelDisplayWeightedDecay(SPL_DECAY_RATE, SPL_WINDOW_TIME)
+                    }
+                }
+                DoubleArray(indicators.nominalFrequencies.size) { index ->
+                    levelDisplayBands?.get(index)
+                        ?.getWeightedValue(indicators.thirdOctave[index])
+                        ?: 0.0
+                }
+            }
     }
 }
