--- composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/components/Greeting.kt
@@ -1,12 +1,0 @@
-package org.noiseplanet.noisecapture.ui.components
-
-import getPlatform
-
-class Greeting {
-
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
